Reverts to an earlier version of deleting programs, teams but removes expensive operations

+++ supervisor.py
@@ -62,28 +62,22 @@ def train(run_id, configuration, num_generations):
         survivor_ids = model.get_survivor_ids(run_id, generation)
 
         for team in model.teamPopulation:
+            if str(team.id) not in [str(team_id) for team_id in root_teams]:
+                continue
             if team.id not in survivor_ids:
                 if team.luckyBreaks > 0:
                     team.luckyBreaks -= 1
-                    teams_to_update_lucky_breaks += team
                     Database.update_team(run_id, team, team.luckyBreaks)
                     continue
 
                 for program in team.programs:
-                    programs_to_remove += (program, team)
+                    Database.remove_program(run_id, program, team)
 
-                teams_to_remove += team
+                Database.remove_team(run_id, team)
 
                 for _team in model.teamPopulation:
                     if str(team.id) == _team.id:
                         model.teamPopulation.remove(team)
-
-        print("Removing programs and teams EXACTLY at this point")
-        for program, team in programs_to_remove:
-            Database.remove_program(run_id, program, team)
-
-        for team in teams_to_remove:
-            Database.remove_team(run_id, team)
 
         print("Cloning existing teams and adding new teams to the database now")
         model.repopulate(run_id, generation)
